+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add status in `ProjectCard`.
+
 ## [11.3.0] - 2017-05-02
 
 Features: