+++ CHANGELOG.md
@@ -4,9 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add status in `ProjectCard`.
-- Fix: Fix tag `Title` in `ProjectCardComponent`.
-- Fix: Fix test launch.
+## [11.4.0] - 2017-05-04
+
+Feature:
+- Add status in `ProjectCard`.
+
+Fixes:
+- Fix `LegoGrid` grid.
+- Fix tag `Title` in `ProjectCardComponent`.
+- Fix test launch.
 
 ## [11.3.0] - 2017-05-02